api/requirements/base.txt
Django==2.0.8
djangorestframework==3.8.2
python-decouple==3.1
requests>=2.20.0
psycopg2-binary==2.7.5
cloudinary==1.14.0
Pillow==5.2.0